add job email with run report

+++ bin/tar.sh
@@ -103,9 +103,9 @@ logDir=$dFolder-log
 
 mkdir -p $dFolder $logDir
 
-[ -f $dFolder/runStarted.lock ] && echo A run was started earlier on the folder. Before re-submitting, please make sure all jobs from that run are finished, then manually delete $dFolder/runStarted.lock"; exit 1; }
+[ -f $logDir/runStarted.lock ] && { echo A run was started earlier on the folder. Before re-submitting, please make sure all jobs from that run are finished, then manually delete $logDir/runStarted.lock; exit 1; }
 
-touch $dFolder/runStarted.lock 
+touch $logDir/runStarted.lock 
 
 rm -r $logDir/exclusive 2>/dev/null || true 
 
@@ -164,7 +164,6 @@ elif [[ "$action" == sbatch ]]; then
     echo "echo \$SLURM_ARRAY_TASK_ID start time \$(date) \$SLURM_JOBID >> $logDir/runTime.txt" >> $logDir/array.sh
     echo dFolder=$dFolder >> $logDir/array.sh
     echo logDir=$logDir >> $logDir/array.sh
-    echo pwd=\$PWD >> $logDir/array.sh
 
     echo "start_row=\$(( (SLURM_ARRAY_TASK_ID - 1) * $rows_per_job + 1 ))" >> $logDir/array.sh 
     echo "end_row=\$(( SLURM_ARRAY_TASK_ID * $rows_per_job ))"  >> $logDir/array.sh 
@@ -180,9 +179,9 @@ elif [[ "$action" == sbatch ]]; then
     echo done >> $logDir/array.sh 
     echo echo done >> $logDir/array.sh
     echo "echo \$SLURM_ARRAY_TASK_ID end time \$(date) \$SLURM_JOBID >> $logDir/runTime.txt" >> $logDir/array.sh
-    echo "cd $pwd" >> $logDir/array.sh
-    echo "summarizeRun.sh | mail -s "\${dFolder##*/}/\$SLURM_ARRAY_TASK_ID is done\" $USER" >> $logDir/array.sh
-    
+
+    echo "echo -e \"Subject: ${dFolder##*/}/\$SLURM_ARRAY_TASK_ID is done\n\`summarizeRun.sh $dFolder\`\" | sendmail `head -n 1 ~/.forward` " >> $logDir/array.sh
+
     echo Slurm script:
     cat $logDir/array.sh
     
@@ -225,7 +224,7 @@ elif [[ "$action" == esbatch ]]; then
     
     echo "echo \$jIndex end time \$(date) \$SLURM_JOBID >> $logDir/runTime.txt" >> $logDir/job.sh 
     
-    echo "summarizeRun.sh | mail -s "\${dFolder##*/}/\$jIndex/\$SLURM_JOBID is done\" $USER" >> $logDir/job.sh
+    echo "echo -e \"Subject: ${dFolder##*/}/\$jIndex/\$SLURM_JOBID is done\n\`summarizeRun.sh $dFolder\`\" | sendmail `head -n 1 ~/.forward` " >> $logDir/job.sh
     
     echo "while ! mkdir $logDir/exclusive 2>/dev/null; do" >> $logDir/job.sh 
     echo "  sleep \$((1 + RANDOM % 10))" >> $logDir/job.sh
